feat(settings): promote Notifications from settings accordion to its own tab
Notifications was buried at the bottom of the Settings accordion list, one
more click away than Panels/Sources. Since the feature is Pro-gated and
channel-heavy (Telegram pairing, Slack/Discord OAuth, webhook URL entry,
quiet hours, digest scheduling), a dedicated tab gives it the surface it
needs and makes the upsell visible to free users.

Extracts the notifications HTML and attach() logic into
src/services/notifications-settings.ts. Removes the inline block and now
unused imports (notification-channels, clerk, entitlements.hasTier,
variant, uqr, QUIET_HOURS/DIGEST_CRON rollout flags) from
preferences-content.ts. Adds a 'notifications' tab between Sources and
API Keys in UnifiedSettings (web only; desktop app keeps the old layout).

Rollout-flag tests (digest/quiet-hours) now read from the new module.
