+++ src/locales/en.json
@@ -221,6 +221,7 @@
     "tabSettings": "Settings",
     "tabPanels": "Panels",
     "tabSources": "Sources",
+    "tabNotifications": "Notifications",
     "languageLabel": "Language",
     "sourceRegionAll": "All",
     "sourceRegionWorldwide": "Worldwide",
